feat(1d): update meta validation rule for activity answer creation
